stage1_ultimate/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "streetvision-stage1"
version = "1.0.0"
description = "StreetVision Training Pipeline - Stage 1 Ultimate (Python 3.14+ BLEEDING EDGE)"
readme = "README.md"
requires-python = ">=3.14"  # Python 3.14+ - BLEEDING EDGE with improved async, better performance!
license = {text = "MIT"}
authors = [
    {name = "StreetVision Team"},
]
keywords = ["computer-vision", "roadwork", "classification", "dinov2", "pytorch-lightning", "research", "mining"]

dependencies = [
    # LATEST PyTorch with Flash Attention 3 support
    "torch>=2.5.0",
    "torchvision>=0.20.0",

    # LATEST Lightning
    "lightning>=2.4.0",

    # LATEST Hydra
    "hydra-core>=1.3.0",

    # LATEST Vision/ML libraries
    "timm>=1.0.0",
    "transformers>=4.40.0",
    "peft>=0.13.0",  # Latest PEFT for ExPLoRA

    # LATEST Scientific computing
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "numpy>=2.0.0",  # NumPy 2.0+ for speed

    # Utilities
    "pillow>=11.0.0",
    "tqdm>=4.66.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "black>=24.2.0",
    "isort>=5.13.0",
    "pre-commit>=3.6.0",
]

calibration = [
    "netcal>=1.3.0",
    "uncertainty-toolbox>=0.1.1",
]

deployment = [
    "onnx>=1.16.0",
    "onnxruntime>=1.17.0",
    "tensorrt>=10.0.0",
]

mlops = [
    "dvc>=3.48.0",
    "wandb>=0.16.0",
    "mlflow>=2.11.0",
]

all = [
    "streetvision-stage1[dev,calibration,deployment,mlops]",
]

[project.scripts]
streetvision-train = "scripts.train_cli:main"

[project.urls]
Repository = "https://github.com/your-org/streetvision"
Documentation = "https://streetvision.readthedocs.io"

# Ruff configuration (replaces flake8, isort, pydocstyle)
[tool.ruff]
line-length = 100
target-version = "py314"  # Python 3.14 - BLEEDING EDGE!

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

# Black configuration
[tool.black]
line-length = 100
target-version = ["py314"]  # Python 3.14 - BLEEDING EDGE!
include = '\.pyi?$'

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["src"]

# MyPy configuration
[tool.mypy]
python_version = "3.14"  # Python 3.14 - BLEEDING EDGE!
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start lenient, tighten later
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if False:",
    "if TYPE_CHECKING:",
]

# Setuptools configuration
[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]
exclude = ["tests*", "docs*"]
